+++ x/mint/module.go
@@ -160,8 +160,8 @@ func (am AppModule) ExportGenesis(ctx context.Context) (json.RawMessage, error) 
 func (AppModule) ConsensusVersion() uint64 { return ConsensusVersion }
 
 // BeginBlock returns the begin blocker for the mint module.
-func (am AppModule) BeginBlock(ctx context.Context) error {
-	return am.keeper.BeginBlocker(ctx, am.mintFn)
+func (am AppModule) BeginBlock(ctx sdk.Context, _ abci.RequestBeginBlock) {
+	BeginBlocker(ctx, am.keeper)
 }
 
 // AppModuleSimulation functions